+++ molten_jwt/token.py
@@ -7,38 +7,35 @@ from authlib.specs.rfc7519 import jwt
 from authlib.common.errors import AuthlibBaseError
 from molten import Settings
 
-from .exceptions import ConfigurationError, AuthenticationError
+from .exceptions import AuthenticationError, ConfigurationError
 
 logger = logging.getLogger(__name__)
 
 
 # TODO init function should take parameters not a settings dictionary.
 class JWT:
-    """The `JWT` instance is used to both encode and decode JSON Web Tokens
-    (JWTs) within your application. This class requires at a minimum that you
-    provide a `JWT_SECRET` within a `molten.Settings` dictionary"""
+    """The `JWT` class provides the core methods to encode and decode JSON
+    Web Tokens in your application or middleware. All tokens produced are
+    signed with a key and algorithm according to the JSON Web Signature(JWS)
+    specification.
+    """
 
-    def __init__(self, settings: Settings) -> None:
-        self.secret: str = settings.get("JWT_SECRET")
-        self.algorithm: str = settings.get("JWT_ALGORITHM", "HS256")
-        self.authorization_prefix: str = settings.get(
-            "JWT_AUTHORIZATION_PREFIX", "bearer"
-        )
-        self.identity_claim: str = settings.get("JWT_USER_ID", "sub")
-        self.user_name_claim: str = settings.get("JWT_USER_NAME", "name")
-        self.options: Dict = settings.get("JWT_OPTIONS", {})
+    def __init__(self, key: str, alg: str, **options) -> None:
+        self.key = key
+        self.alg = alg
+        self.options = options
+        # self.authorization_prefix: str = settings.get(
+        #     "JWT_AUTHORIZATION_PREFIX", "bearer"
+        # )
 
-        if self.secret is None:
-            raise ConfigurationError(
-                "JWT_SECRET passed as part of settings on instantiation"
-            )
+        # self.options: Dict = settings.get("JWT_OPTIONS", {})
 
     def encode(self, payload: Dict) -> str:
-        """Generates a JWT auth token"""
+        """Generates a JSON Web Token as a utf8 string from a dictionary payload."""
         try:
 
             return jwt.encode(
-                header={"alg": self.algorithm}, payload=payload, key=self.secret
+                header={"alg": self.alg}, payload=payload, key=self.key
             ).decode(encoding="utf8")
 
         except AuthlibBaseError as err:
@@ -46,9 +43,9 @@ class JWT:
 
     # TODO add support for claims to verify and, claim options, and claim params
     def decode(self, token: str) -> Optional[Dict]:
-        """Decodes a JWT auth token"""
+        """Decodes a JWT token"""
         try:
-            payload = jwt.decode(token, self.secret, **self.options)
+            payload = jwt.decode(token, self.key, **self.options)
             if payload == {}:
                 raise AuthenticationError("No payload present in token")
         except BadSignatureError as err:
@@ -62,20 +59,38 @@ class JWT:
         return payload
 
 
-# TODO JWTComponent should be changed to provide a JWT instance for use in JWTIdentity
-class JWTComponent:
-    """A component that sets up a JWT instance for use in encoding
-    and decoding JSON Web Tokens. This component depends on the
-    availability of a `molten.Settings` component.
+def config_jwt_from_settings(settings: Settings) -> JWT:
+    """Configures a `molten_jwt.JWT` instance from a `molten.Settings`
+    instance.
 
-    Your settings dictionary must contain a `JWT_SECRET` setting
+    Your settings dictionary must contain a `JWT_SECRET_KEY` setting
     at a minimum, for use in signing and verifying JWTs. Additionally,
     you may include:
 
     `JWT_ALGORITHM`: Defaults to `HS256`.
+    """
+    key: str = settings.get("JWT_SECRET_KEY")
+    alg: str = settings.get("JWT_ALGORITHM", "HS256")
 
-    `JWT_AUTHORIZATION_PREFIX`: a string for the tokeN scheme. Defaults
-    to `bearer`."""
+    if key is None:
+        raise ConfigurationError(
+            "JWT_SECRET_KEY passed as part of settings on instantiation"
+        )
+
+    return JWT(key=key, alg=alg)
+
+
+class JWTComponent:
+    """A component that configures a single cached JWT instance to be
+    injected through Molten's dependency injection system. This component
+    depends on the availability of a `molten.Settings` component.
+
+    Your settings dictionary must contain a `JWT_SECRET_KEY` setting
+    at a minimum, for use in signing and verifying JWTs. Additionally,
+    you may include:
+
+    `JWT_ALGORITHM`: Defaults to `HS256`.
+    """
 
     is_cacheable = True
     is_singleton = True
@@ -84,4 +99,4 @@ class JWTComponent:
         return parameter.annotation is JWT
 
     def resolve(self, settings: Settings) -> JWT:
-        return JWT(settings)
+        return config_jwt_from_settings(settings)