+++ tests/test_token.py
@@ -2,27 +2,29 @@ import pytest
 import inspect
 
 from molten import Settings
-from molten_jwt.token import JWT, JWTComponent
+from molten_jwt.token import JWT, JWTComponent, config_jwt_from_settings
 from molten_jwt.exceptions import ConfigurationError
 
 secret = "keepthissafe"
 
-settings = Settings({"JWT_SECRET": secret})
+settings = Settings({"JWT_SECRET_KEY": secret})
 
 
-def test_JWT_raises_config_error():
+def test_config_jwt_from_settings_raises_config_error():
     with pytest.raises(ConfigurationError):
-        JWT(settings={})
+        config_jwt_from_settings(settings={})
 
 
+# TODO enumerate testing with the other JWS algorithms
 def test_JWT_encode(testing_token):
-    jwt = JWT(settings)
+    jwt = JWT(key=secret, alg="HS256")
     payload = {"sub": "1234567890", "name": "John Doe", "iat": 1516239022}
     assert testing_token == jwt.encode(payload)
 
 
+# TODO enumerate testing with the other JWS algorithms
 def test_JWT_decode(testing_token):
-    jwt = JWT(settings)
+    jwt = JWT(key="keepthissafe", alg="HS256")
     payload = {"sub": "1234567890", "name": "John Doe", "iat": 1516239022}
     token = jwt.decode(testing_token)
     assert payload.get("sub") == token.get("sub")